Cleaned 2 obsolete entries and reordered the rest
Removed 3 (more or less) Go specific entries that were needed in the
past, but now with the `language-go` package being improved, these are
not needed/used anymore.

So `.support.function.decl`, `.keyword.operator.bracket` and
`.keyword.operator.punctuation` are deleted (they were previously all
added by me specifically for supporting `language-go` better, so it
should be save to remove them again now they aren’t used anymore).

For the rest I reordered the entries to be alphabetical to make it a
little cleaner and easier to find entries.